Fix/double to big decimal

+++ src/main/java/com/vape/data/streaming/model/FFTDataPointModel.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -15,7 +14,7 @@ public class FFTDataPointModel {
     @Id
     private String id;
     private String sensorDataPointId;
-    private List<BigDecimal> imaginaryValue;
-    private List<BigDecimal> realValue;
+    private List<Double> imaginaryValue;
+    private List<Double> realValue;
     private String timestamp;
 }
